+++ Editor/Playa/Renderer/Table/TableContentElement.cs
@@ -66,6 +66,7 @@ namespace SaintsField.Editor.Playa.Renderer.Table
 
             // Filled
             _tableContentContainer = new VisualElement();
+            _tableContentContainer.AddToClassList("unity-list-view__scroll-view--with-footer");
             Add(_tableContentContainer);
 
             this.TrackPropertyValue(fieldWithInfo.SerializedProperty, _ => OnArrayPropertyChanged());