+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## WIP ##
+
+1.  Unity Fix: When nested list view, only one element will be styled as focused, fix [#279](https://github.com/TylerTemp/SaintsField/issues/279)
+2.  Fix: `Table` with all footer buttons hidden, there was an empty container at the footer
+3.  Improve: `ListDrawerSettings` now always have the border with a total number field, just like Unity's default list drawer.
+4.  Improve: `Table` content now styled the say way as Unity's default list drawer
+
 ## 5.18.4 ##
 
 1.  Improve: When in Horizental layout, if the target drawer is not from SaintsField, and it does return a `BaseField<>` type, the drawer will be break into 2 rows. This effects horizental layout, `SaintsDictionary`, `Table`. [#400](https://github.com/TylerTemp/SaintsField/issues/400)